server/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
mysql-connector-python==8.1.0
openai==1.9.2
python-dotenv==1.0.0
